+++ sdk/core/azure-core/src/main/java/com/azure/core/util/tracing/Tracer.java
@@ -248,7 +248,28 @@ public interface Tracer {
      * @param attributes the additional attributes to be set for the event.
      * @param timestamp The instant, in UTC, at which the event will be associated to the span.
      * @throws NullPointerException if {@code eventName} is {@code null}.
+     * @deprecated Use {@link #addEvent(String, Map, OffsetDateTime, Context)}
      */
+    @Deprecated
     default void addEvent(String name, Map<String, Object> attributes, OffsetDateTime timestamp) {
+    }
+
+    /**
+     * Adds an event to the span present in the {@code Context} with the provided {@code timestamp}
+     * and {@code attributes}.
+     * <p>This API does not provide any normalization if provided timestamps are out of range of the current
+     * span timeline</p>
+     * <p>Supported attribute values include String, double, boolean, long, String [], double [], long [].
+     * Any other Object value type and null values will be silently ignored.</p>
+     *
+     * @param name the name of the event.
+     * @param attributes the additional attributes to be set for the event.
+     * @param timestamp The instant, in UTC, at which the event will be associated to the span.
+     * @param context the call metadata containing information of the span to which the event should be associated with.
+     * @throws NullPointerException if {@code eventName} is {@code null}.
+     */
+    default void addEvent(String name, Map<String, Object> attributes, OffsetDateTime timestamp,
+                          Context context) {
+
     }
 }