+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/implementation/FeedResponseDiagnostics.java
@@ -33,7 +33,7 @@ public class FeedResponseDiagnostics {
         this.clientSideRequestStatisticsList = Collections.synchronizedList(new ArrayList<>());
     }
 
-    Map<String, QueryMetrics> getQueryMetricsMap() {
+    public Map<String, QueryMetrics> getQueryMetricsMap() {
         return queryMetricsMap;
     }
 
@@ -65,6 +65,17 @@ public class FeedResponseDiagnostics {
                     .append(EQUALS)
                     .append(Duration.between(diagnosticsContext.getStartTimeUTC(),
                         diagnosticsContext.getEndTimeUTC()).toMillis()).append(System.lineSeparator());
+                if (diagnosticsContext.getRequestTimeline() != null) {
+                    try {
+                        stringBuilder.append(QUERY_PLAN + SPACE + "RequestTimeline ")
+                            .append(EQUALS)
+                            .append(Utils.getSimpleObjectMapper().writeValueAsString(diagnosticsContext.getRequestTimeline()))
+                            .append(System.lineSeparator())
+                            .append(System.lineSeparator());
+                    } catch (JsonProcessingException e) {
+                        LOGGER.error("Error while parsing diagnostics ", e);
+                    }
+                }
             }
         }
 
@@ -86,6 +97,10 @@ public class FeedResponseDiagnostics {
 
     public void setDiagnosticsContext(QueryInfo.QueryPlanDiagnosticsContext diagnosticsContext) {
         this.diagnosticsContext = diagnosticsContext;
+    }
+
+    public QueryInfo.QueryPlanDiagnosticsContext getQueryPlanDiagnosticsContext() {
+        return diagnosticsContext;
     }
 
     /**