+++ sdk/keyvault/azure-security-keyvault-jca/src/test/java/com/azure/security/keyvault/jca/KeyVaultCertificatesTest.java
@@ -26,13 +26,12 @@ public class KeyVaultCertificatesTest {
 
     @BeforeEach
     public void beforeEach() {
-
         List<String> aliases = new ArrayList<>();
         aliases.add("myalias");
         when(keyVaultClient.getAliases()).thenReturn(aliases);
         when(keyVaultClient.getKey("myalias", null)).thenReturn(key);
         when(keyVaultClient.getCertificate("myalias")).thenReturn(certificate);
-        keyVaultCertificates = new KeyVaultCertificates(0, keyVaultClient);
+        keyVaultCertificates = new KeyVaultCertificates(60_000, keyVaultClient);
     }
 
     @Test
@@ -51,11 +50,12 @@ public class KeyVaultCertificatesTest {
     }
 
     @Test
-    public void testRefreshAndGetAliasByCertificate() throws InterruptedException {
+    public void testRefreshAndGetAliasByCertificate() {
         Assertions.assertEquals(keyVaultCertificates.refreshAndGetAliasByCertificate(certificate), "myalias");
+        Assertions.assertEquals(keyVaultCertificates.getCertificates().get("myalias"), certificate);
         when(keyVaultClient.getAliases()).thenReturn(null);
-        Thread.sleep(10); // TODO: (rujche) update this logic
         Assertions.assertNotEquals(keyVaultCertificates.refreshAndGetAliasByCertificate(certificate), "myalias");
+        Assertions.assertNull(keyVaultCertificates.getCertificates().get("myalias"));
     }
 
     @Test
@@ -63,21 +63,6 @@ public class KeyVaultCertificatesTest {
         Assertions.assertTrue(keyVaultCertificates.getAliases().contains("myalias"));
         keyVaultCertificates.deleteEntry("myalias");
         Assertions.assertFalse(keyVaultCertificates.getAliases().contains("myalias"));
-    }
-
-    @Test
-    public void testCertificatesNeedRefresh() throws InterruptedException {
-        keyVaultCertificates = new KeyVaultCertificates(1000, keyVaultClient);
-        Assertions.assertTrue(keyVaultCertificates.certificatesNeedRefresh());
-        keyVaultCertificates.getAliases();
-        Assertions.assertFalse(keyVaultCertificates.certificatesNeedRefresh());
-        Thread.sleep(10);
-        KeyVaultCertificates.updateLastForceRefreshTime();
-        Assertions.assertTrue(keyVaultCertificates.certificatesNeedRefresh());
-        keyVaultCertificates.getAliases();
-        Assertions.assertFalse(keyVaultCertificates.certificatesNeedRefresh());
-        Thread.sleep(2000);
-        Assertions.assertTrue(keyVaultCertificates.certificatesNeedRefresh());
     }
 
 }