+++ sdk/metricsadvisor/azure-ai-metricsadvisor/src/main/java/com/azure/ai/metricsadvisor/administration/models/MetricAlertConfiguration.java
@@ -61,7 +61,7 @@ public final class MetricAlertConfiguration {
      *
      * @return The negation operation value.
      */
-    public boolean isNegationOperationEnabled() {
+    public Boolean isNegationOperationEnabled() {
         return this.negationOperation;
     }