+++ sdk/communication/azure-communication-common/CHANGELOG.md
@@ -3,6 +3,9 @@
 ## 1.1.0-beta.1 (Unreleased)
 
 
+## 1.0.3 (2021-06-28)
+Updated `azure-communication-common` version
+
 ## 1.0.2 (2021-06-09)
 Updated `azure-communication-common` version
 
@@ -51,4 +54,4 @@ This package contains common code for Azure Communication Service libraries. For
 This is a Public Preview version, so breaking changes are possible in subsequent releases as we improve the product. To provide feedback, please submit an issue in our [Azure SDK for Java GitHub repo](https://github.com/Azure/azure-sdk-for-java/issues).
 
 <!-- LINKS -->
-[read_me]: https://github.com/Azure/azure-sdk-for-java/blob/master/sdk/communication/azure-communication-common/README.md
+[read_me]: https://github.com/Azure/azure-sdk-for-java/blob/main/sdk/communication/azure-communication-common/README.md