+++ sdk/signalr/azure-resourcemanager-signalr/CHANGELOG.md
@@ -1,7 +1,35 @@
 # Release History
 
-## 1.0.0-beta.2 (Unreleased)
+## 1.0.0-beta.3 (Unreleased)
 
+
+## 1.0.0-beta.2 (2021-07-09)
+
+- Azure Resource Manager SignalR client library for Java. This package contains Microsoft Azure SDK for SignalR Management SDK. REST API for Azure SignalR Service. Package tag package-2021-06-01-preview. For documentation on how to use this package, please see [Azure Management Libraries for Java](https://aka.ms/azsdk/java/mgmt).
+
+### Features Added
+
+#### `SignalRManager$Configurable` was modified
+
+* `withScope(java.lang.String)` was added
+
+#### `models.SignalRResource$Definition` was modified
+
+* `withDisableLocalAuth(java.lang.Boolean)` was added
+* `withPublicNetworkAccess(java.lang.String)` was added
+* `withDisableAadAuth(java.lang.Boolean)` was added
+
+#### `models.SignalRResource` was modified
+
+* `disableLocalAuth()` was added
+* `publicNetworkAccess()` was added
+* `disableAadAuth()` was added
+
+#### `models.SignalRResource$Update` was modified
+
+* `withDisableAadAuth(java.lang.Boolean)` was added
+* `withPublicNetworkAccess(java.lang.String)` was added
+* `withDisableLocalAuth(java.lang.Boolean)` was added
 
 ## 1.0.0-beta.1 (2021-04-14)